appcode/backend/my-app/requirements.txt
azure-functions
fastapi
uvicorn
azure-cosmos
azure-identity